requirements.txt
Flask==0.11.1
gunicorn==19.6.0
google-cloud-storage==0.20.0
google-cloud-datastore
jinja2>=2.6
twilio

